fix ut bugs
